requirements.txt
streamlit
python-doctr
python-doctr[viz]
matplotlib
Pillow
